Htpm 175
checked

+++ client/src/redux/cartReducer/cartActions.js
@@ -3,13 +3,16 @@ export const ADD_PRODUCT = "ADD_PRODUCT";
 export const DELETE_PRODUCT = "DELETE_PRODUCT";
 export const TOTAL = "TOTAL";
 export const INCREMENTQ = "INCREMENTQ";
-
-var userId = 1;
+export const USERLOGGED = "USERLOGGED";
+export const EMPTY = "EMPTY";
 
 export function addProduct(product) {
-  axios.post(`http://localhost:3001/cart/${userId}/${product.id}`, {
-    productId: product.id,
-  });
+  const userId = localStorage.getItem("user");
+  if (userId !== "0") {
+    axios.post(`http://localhost:3001/cart/${userId}/${product.id}`, {
+      productId: product.id,
+    });
+  }
   return function (dispatch) {
     dispatch({
       type: "ADD_PRODUCT",
@@ -19,7 +22,10 @@ export function addProduct(product) {
 }
 
 export function deleteProduct(product) {
-  axios.delete(`http://localhost:3001/cart/${userId}/${product.id}`);
+  const userId = localStorage.getItem("user");
+  if (userId !== "0") {
+    axios.delete(`http://localhost:3001/cart/${userId}/${product.id}`);
+  }
   return function (dispatch) {
     dispatch({
       type: "DELETE_PRODUCT",
@@ -36,11 +42,14 @@ export function totalPrice() {
   };
 }
 export function incrementQ(product, value) {
-  axios.put(`http://localhost:3001/cart/user/${product.id}`, {
-    id: product.id,
-    quantity: value,
-    userId: userId,
-  });
+  const userId = localStorage.getItem("user");
+  if (userId !== "0") {
+    axios.put(`http://localhost:3001/cart/user/${product.id}`, {
+      id: product.id,
+      quantity: value,
+      userId: userId,
+    });
+  }
   return function (dispatch) {
     dispatch({
       type: "INCREMENTQ",
@@ -48,6 +57,34 @@ export function incrementQ(product, value) {
         product,
         value,
       },
+    });
+  };
+}
+export function userLogged(cart) {
+  return function (dispatch) {
+    dispatch({
+      type: "USERLOGGED",
+      payload: cart,
+    });
+  };
+}
+
+export function emptyDb() {
+  const userId = localStorage.getItem("user");
+  if (userId !== "0") {
+    axios.delete(`http://localhost:3001/cart/${userId}/items/delete`);
+  }
+  return function (dispatch) {
+    dispatch({
+      type: "EMPTY",
+    });
+  };
+}
+
+export function emptyCart() {
+  return function (dispatch) {
+    dispatch({
+      type: "EMPTY",
     });
   };
 }
